+++ NEWS.md
@@ -1,6 +1,6 @@
 **If you are viewing this file on CRAN, please check [latest news on GitHub](https://github.com/psychbruce/bruceR/blob/master/NEWS.md) where the formatting is also better.**
 
-# bruceR 0.7.0 (in development)
+# bruceR 0.7.0 (May 2021)
 
 ### New Features